Created a DrawerContent folder and sperated DrawerContent components into DrawerSection & UserInfo components
