+++ Components/Navigators/DrawerNaviagator.js
@@ -3,8 +3,8 @@ import React from 'react';
 import {createDrawerNavigator} from '@react-navigation/drawer';
 
 import StackScreen from './StackNavigator';
-import DrawerContent from '../Screens/DrawerContent';
-import ProfileScreen from '../Screens/ProfileScreen';
+import DrawerContent from '../Screens/DrawerContent/DrawerContent';
+
 import BookMarkScreen from '../Screens/BookMarks';
 import SettingsScreen from '../Screens/Settings';
 import SupportScreen from '../Screens/Support';
@@ -14,8 +14,7 @@ const Drawer = createDrawerNavigator();
 const DrawerScreen = () => {
   return (
     <Drawer.Navigator drawerContent={(props) => <DrawerContent {...props} />}>
-      <Drawer.Screen name="Home" component={StackScreen} />
-      <Drawer.Screen name="Profile" component={ProfileScreen} />
+      <Drawer.Screen name="HomeDrawer" component={StackScreen} />
       <Drawer.Screen name="BookMarks" component={BookMarkScreen} />
       <Drawer.Screen name="Settings" component={SettingsScreen} />
       <Drawer.Screen name="Support" component={SupportScreen} />
