--- Components/Screens/DrawerContent.js
@@ -1,180 +1,0 @@
-import React, {useState} from 'react';
-
-import {View, StyleSheet} from 'react-native';
-
-import {DrawerContentScrollView, DrawerItem} from '@react-navigation/drawer';
-
-import {
-  Avatar,
-  Text,
-  Title,
-  Caption,
-  Paragraph,
-  Drawer,
-  TouchableRipple,
-  Switch,
-  useTheme,
-} from 'react-native-paper';
-
-import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-Icon.loadFont();
-
-const DrawerContent = (props) => {
-  const [isDarkTheme, setIsDarkTheme] = useState(false);
-  const toggleTheme = () => setIsDarkTheme(!isDarkTheme);
-  return (
-    <View style={styles.drawerContent}>
-      <DrawerContentScrollView {...props}>
-        <View style={styles.drawerContent}>
-          <View style={styles.userInfo}>
-            <View style={styles.userDetails}>
-              <Avatar.Image
-                source={require('../../assests/myimage.png')}
-                size={50}
-              />
-              <View style={styles.user}>
-                <Title style={styles.title}>Vimal Raveendran</Title>
-                <Caption style={styles.caption}>React Developer</Caption>
-              </View>
-            </View>
-            <View style={styles.subInfo}>
-              <View style={styles.subSection}>
-                <Paragraph style={[styles.paragraph, styles.caption]}>
-                  80
-                </Paragraph>
-                <Caption style={styles.caption}>Following</Caption>
-              </View>
-              <View style={styles.subSection}>
-                <Paragraph style={[styles.paragraph, styles.caption]}>
-                  100
-                </Paragraph>
-                <Caption style={styles.caption}>Followers</Caption>
-              </View>
-            </View>
-          </View>
-          <Drawer.Section style={styles.drawerSection}>
-            <DrawerItem
-              icon={({color, size}) => (
-                <Icon name="home-outline" color={color} size={size} />
-              )}
-              label="Home"
-              onPress={() => {
-                props.navigation.navigate('Home');
-              }}
-            />
-            <DrawerItem
-              icon={({color, size}) => (
-                <Icon name="account-outline" color={color} size={size} />
-              )}
-              label="Profile"
-              onPress={() => {
-                props.navigation.navigate('Profile');
-              }}
-            />
-            <DrawerItem
-              icon={({color, size}) => (
-                <Icon name="bookmark-outline" color={color} size={size} />
-              )}
-              label="Bookmarks"
-              onPress={() => {
-                props.navigation.navigate('BookMarks');
-              }}
-            />
-            <DrawerItem
-              icon={({color, size}) => (
-                <Icon name="eye-settings-outline" color={color} size={size} />
-              )}
-              label="Settings"
-              onPress={() => {
-                props.navigation.navigate('Settings');
-              }}
-            />
-            <DrawerItem
-              icon={({color, size}) => (
-                <Icon name="account-check-outline" color={color} size={size} />
-              )}
-              label="Support"
-              onPress={() => {
-                props.navigation.navigate('Support');
-              }}
-            />
-          </Drawer.Section>
-          <Drawer.Section title="Preference">
-            <TouchableRipple
-              onPress={() => {
-                toggleTheme();
-              }}>
-              <View style={styles.preference}>
-                <Text>Dark Theme</Text>
-                <View pointerEvents="none">
-                  <Switch value={isDarkTheme} />
-                </View>
-              </View>
-            </TouchableRipple>
-          </Drawer.Section>
-        </View>
-      </DrawerContentScrollView>
-      <Drawer.Section style={styles.bottomDrawer}>
-        <DrawerItem
-          icon={({color, size}) => (
-            <Icon name="exit-to-app" color={color} size={size} />
-          )}
-          label="Sign-Out"
-        />
-      </Drawer.Section>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  drawerContent: {
-    flex: 1,
-  },
-  userInfo: {
-    paddingLeft: 20,
-  },
-  title: {
-    fontSize: 16,
-    marginTop: 3,
-    fontWeight: '700',
-  },
-  userDetails: {
-    flexDirection: 'row',
-    marginTop: 15,
-  },
-  user: {
-    marginLeft: 15,
-  },
-  caption: {
-    fontSize: 14,
-    lineHeight: 14,
-  },
-  subInfo: {
-    flexDirection: 'row',
-    marginTop: 20,
-  },
-  paragraph: {
-    fontWeight: 'bold',
-    marginRight: 3,
-  },
-
-  subSection: {
-    flexDirection: 'row',
-    paddingLeft: 20,
-  },
-  drawerSection: {
-    marginTop: 15,
-  },
-  preference: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginHorizontal: 14,
-  },
-  bottomDrawer: {
-    marginBottom: 15,
-    borderTopColor: '#f4f4f4',
-    borderTopWidth: 1,
-  },
-});
-export default DrawerContent;
